fix(services): convert core services
